+++ src/main/java/controller/UserListController.java
@@ -9,12 +9,14 @@ public class UserListController extends AbstractController {
     private static final String NEED_LOGIN_MESSAGE = "로그인이 필요합니다.";
 
     @Override
-    public void doPost(HttpRequest httpRequest, HttpResponse httpResponse) {
-    }
-
-    @Override
     public void doGet(HttpRequest httpRequest, HttpResponse httpResponse) {
-        validateLogin(httpRequest);
+        try {
+            validateLogin(httpRequest);
+        } catch (UserAuthorizationException e) {
+            logger.error(e.getMessage());
+            httpResponse.sendRedirect("/user/login.html");
+            return;
+        }
         byte[] page = ResourceLoader.getDynamicPageWithUser("user/list");
         logger.debug("Succeeded in loading user-list page");
         httpResponse.forwardBody(page);
@@ -23,7 +25,6 @@ public class UserListController extends AbstractController {
     private void validateLogin(HttpRequest httpRequest) {
         String cookie = httpRequest.getHeader("Cookie");
         if (!isLogin(cookie)) {
-            logger.error(NEED_LOGIN_MESSAGE);
             throw new UserAuthorizationException(NEED_LOGIN_MESSAGE);
         }
     }
