+++ src/main/java/webserver/HttpStatusCode.java
@@ -6,7 +6,8 @@ public enum HttpStatusCode {
     FOUND(302, "Found"),
     BAD_REQUEST(400, "Bad Request"),
     UNAUTHORIZED(401, "Unauthorized"),
-    NOT_FOUND(404, "Not Found");
+    NOT_FOUND(404, "Not Found"),
+    METHOD_NOT_ALLOWED(405, "Method Not Allowed");
 
     private int code;
     private String description;
